refactor components into its own file
