+++ src/App.jsx
@@ -4,9 +4,7 @@ import GamePanel from './components/GamePanel'
 function App() {
 
   return (
-    <>
       <GamePanel/>
-    </>
   )
 }
