+++ src/components/GamePanel.css
@@ -1,32 +1,3 @@
-#heart { 
-  position: relative; 
-  width: 100px; 
-  height: 90px; 
-} 
-
-#heart:before, #heart:after { 
-  position: absolute; 
-  content: ""; 
-  left: 50px; 
-  top: 0; 
-  width: 50px; 
-  height: 80px; 
-  background: red; 
-  border-radius: 50px 50px 0 0; 
-  transform: rotate(-45deg); 
-  transform-origin: 0 100%; 
-} 
-
-#heart:after { 
-  left: 0; 
-  transform: rotate(45deg); 
-  transform-origin :100% 100%; 
-}
-
-.heart-container {
-  display: flex;
-  justify-content: center;
-}
 
 .question {
   font-family: cursive;
